two custom events are called based on function
remove() method is invoked on 'remove-event' custom event, this then calls a separate method to find the index to avoid code duplication.

closeForm() is invoked on 'close-form' custom event, this just closes the form using a boolean and sets the event Object back to empty.

findEventIndex() is invoked to find the index using the id it is passed and linearly searches the array and returns the found index, otherwise returns -1.